refactor: extract analyzer widgets
